requirements.txt
langchain-openai
langgraph
pandas
tavily-python
pydantic
playwright
bs4
asyncio
python-dotenv